requirements.txt
appnope==0.1.0
decorator==4.0.11
Django==1.8.2
django-haystack==2.6.0
ipython==5.2.2
ipython-genutils==0.1.0
jieba==0.38
Markdown==2.6.8
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==3.4.0
PyMySQL==0.7.10
redis==2.10.5
simplegeneric==0.8.1
traitlets==4.3.1
wcwidth==0.1.7
Whoosh==2.7.4
